// pdf-proxy/types/generated.tsx
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Assignment = {
  __typename?: 'Assignment';
  fid: Scalars['ID'];
  id: Scalars['ID'];
  pages?: Maybe<Array<Page>>;
  name: Scalars['String'];
};

export type Folder = {
  __typename?: 'Folder';
  fid: Scalars['ID'];
  name: Scalars['String'];
  pages?: Maybe<Array<Page>>;
};

export type FolderComponent = {
  id: Scalars['ID'];
  title: Scalars['String'];
};
// ...
export type TableCell = {
  __typename?: 'TableCell';
  row: Scalars['Int'];
  column: Scalars['Int'];
  response?: Maybe<Scalars['String']>;
};

export type TableCellInput = {
  row: Scalars['Int'];
  column: Scalars['Int'];
  response?: Maybe<Scalars['String']>;
};

export type TableComponent = FolderComponent & {
  __typename?: 'TableComponent';
  id: Scalars['ID'];
  title: Scalars['String'];
  rows: Array<Scalars['String']>;
  columns: Array<Scalars['String']>;
  cells?: Maybe<Array<TableCell>>;
};